+++ HISTORY.md
@@ -8,6 +8,7 @@ v1.6.2 ()
 
 * Fixed ubus failure on method not found case
 * Support disable Symlink via FreeIOE
+* Support insecure tls mqtt connection
 
 
 v1.6.1 (2019-08-29)